Wip: cleanup

+++ far2l/PluginW.cpp
@@ -182,7 +182,7 @@ static size_t WINAPI FarKeyToName(int Key,wchar_t *KeyText,size_t Size)
 int WINAPI KeyNameToKeyW(const wchar_t *Name)
 {
 	FARString strN(Name);
-	return KeyNameToKey(strN);
+	return (int)KeyNameToKey(strN);
 }
 
 PluginW::PluginW(PluginManager *owner, const wchar_t *lpwszModuleName):
